+++ src/Illuminate/Database/Query/Builder.php
@@ -2102,7 +2102,7 @@ class Builder
         $property = $this->unions ? 'unionLimit' : 'limit';
 
         if ($value >= 0) {
-            $this->$property = (int) $value;
+            $this->$property = ! is_null($value) ? (int) $value : null;
         }
 
         return $this;
