+++ src/Illuminate/Foundation/Application.php
@@ -33,7 +33,7 @@ class Application extends Container implements ApplicationContract, CachesConfig
      *
      * @var string
      */
-    const VERSION = '8.71.0';
+    const VERSION = '8.72.0';
 
     /**
      * The base path for the Laravel installation.
