+++ src/Illuminate/Foundation/Testing/RefreshDatabase.php
@@ -3,9 +3,12 @@
 namespace Illuminate\Foundation\Testing;
 
 use Illuminate\Contracts\Console\Kernel;
+use Illuminate\Foundation\Testing\Traits\CanConfigureMigrationCommands;
 
 trait RefreshDatabase
 {
+    use CanConfigureMigrationCommands;
+
     /**
      * Define hooks to migrate the database before and after each test.
      *
@@ -73,24 +76,6 @@ trait RefreshDatabase
     }
 
     /**
-     * The parameters that should be used when running "migrate:fresh".
-     *
-     * @return array
-     */
-    protected function migrateFreshUsing()
-    {
-        $seeder = $this->seeder();
-
-        return array_merge(
-            [
-                '--drop-views' => $this->shouldDropViews(),
-                '--drop-types' => $this->shouldDropTypes(),
-            ],
-            $seeder ? ['--seeder' => $seeder] : ['--seed' => $this->shouldSeed()]
-        );
-    }
-
-    /**
      * Begin a database transaction on the testing database.
      *
      * @return void
@@ -130,45 +115,5 @@ trait RefreshDatabase
     {
         return property_exists($this, 'connectionsToTransact')
                             ? $this->connectionsToTransact : [null];
-    }
-
-    /**
-     * Determine if views should be dropped when refreshing the database.
-     *
-     * @return bool
-     */
-    protected function shouldDropViews()
-    {
-        return property_exists($this, 'dropViews') ? $this->dropViews : false;
-    }
-
-    /**
-     * Determine if types should be dropped when refreshing the database.
-     *
-     * @return bool
-     */
-    protected function shouldDropTypes()
-    {
-        return property_exists($this, 'dropTypes') ? $this->dropTypes : false;
-    }
-
-    /**
-     * Determine if the seed task should be run when refreshing the database.
-     *
-     * @return bool
-     */
-    protected function shouldSeed()
-    {
-        return property_exists($this, 'seed') ? $this->seed : false;
-    }
-
-    /**
-     * Determine the specific seeder class that should be used when refreshing the database.
-     *
-     * @return mixed
-     */
-    protected function seeder()
-    {
-        return property_exists($this, 'seeder') ? $this->seeder : false;
     }
 }
